+++ cpp/src/ComplexNumber_io.cpp
@@ -1,29 +1,30 @@
 #include <cmath>
 #include <cstdio>
+#include <iomanip>
 #include "ComplexNumber.h"
 #include "ComplexNumber_io.h"
+using namespace std;
 
 ZMAT_BEGIN_NAMESPACE
 
-void input(ComplexNumber &num) {
+istream& operator >>(istream &in, ComplexNumber &num) {
     double real, image;
-    scanf("%lf %lf", &real, &image);
+    in >> real >> image;
     num.setValue(real, image);
+    return in;
 }
 
-void output(const ComplexNumber &num) {
-    printf("%.6lf ", num.real());
-    if (num.image() <= 0.0) {
-        printf("- ");
-    } else {
-        printf("+ ");
+ostream& operator <<(ostream &out, const ComplexNumber &num) {
+    out << num.real();
+    if (fabs(num.image()) > ZMAT_EPS) {
+        if (num.image() <= 0.0) {
+            out << " - ";
+        } else {
+            out << " + ";
+        }
+        out << num.image() << " i";
     }
-    printf("%.6lf i", fabs(num.image()));
-}
-
-void outputLine(const ComplexNumber &num) {
-    output(num);
-    putchar('\n');
+    return out;
 }
 
 ZMAT_END_NAMESPACE